Adds a configurable "receiver" tag along each influxdb (#54)
By default the receiver tag is an empty string, in which case the
values will not be tagged. If configured (by the "receiver" property
int ruuvi-collector.properties file), then each value sent to influxdb
will have the value in the "receiver" tag.

This allows distinguishing values when the same database is being used
by multiple receivers and allows analyzing packet loss and RSSI
differences, and eventually forming a black/whitelist to minimize
duplicate data in the backing database.