Ignore lescan service stdout
The lescan service produces a lot of log to the journal, as `hcitool lescan` dumps each advert to the standard output. This is suboptimal for e.g. a Raspberry Pi, which is often running off an SD card, where excessive writes are undesirable.

As RuuviCollector does not use the output (lescan just needs to be running so that `hcitool dump` is updated), this commit changes the lescan.service unit to redirect the service's stdout to null. stderr is still written to journal.